+++ lib/analytics/invoice_analytics.rb
