+++ lib/analytics/item_analytics.rb
