+++ lib/analytics/merchant_analytics.rb
