+++ lib/repositories/item_repository.rb
@@ -1,4 +1,4 @@
-require_relative '../lib/item'
+require_relative '../item'
 require_relative 'repository'
 
 # item repo class
