+++ lib/sales_engine.rb
@@ -2,13 +2,13 @@ require 'bigdecimal'
 require 'csv'
 require 'time'
 require_relative 'fileio'
-require_relative 'customer_repository'
-require_relative 'invoice_item_repository'
-require_relative 'invoice_repository'
-require_relative 'item_repository'
-require_relative 'merchant_repository'
-require_relative 'transaction_repository'
-require_relative 'repository'
+require_relative 'repositories/customer_repository'
+require_relative 'repositories/invoice_item_repository'
+require_relative 'repositories/invoice_repository'
+require_relative 'repositories/item_repository'
+require_relative 'repositories/merchant_repository'
+require_relative 'repositories/transaction_repository'
+require_relative 'repositories/repository'
 require_relative 'sales_analyst'
 
 # Sales Engine class for managing data
