+++ test/merchant_analytics_test.rb
@@ -2,7 +2,7 @@ require 'minitest/autorun'
 require 'minitest/pride'
 require_relative '../lib/sales_engine'
 require_relative '../lib/sales_analyst'
-require_relative '../lib/merchant_analytics.rb'
+require_relative '../lib/analytics/merchant_analytics.rb'
 
 class MerchantAnalyticsTest < Minitest::Test
