+++ test/merchant_repository_test.rb
@@ -1,6 +1,6 @@
 require 'time'
 require_relative 'test_helper'
-require_relative '../lib/merchant_repository'
+require_relative '../lib/repositories/merchant_repository'
 require_relative '../lib/fileio'
 
 # merchant repository class
